# api/openapi.yaml
openapi: 3.0.0
info:
  title: Go OpenAPI Starter
  version: "1.0.0"

servers:
  - url: http://localhost:8080

paths:
  /health:
    get:
      summary: Health check
      description: Returns service health status
      operationId: GetHealth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /users:
    get:
      summary: List users
      description: Get all users
      operationId: GetUsers
      responses:
        '200':
          description: A JSON array of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Create user
      description: Create a new user
      operationId: PostUsers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: Invalid input

  /users/{id}:
    get:
      summary: Get user by ID
      operationId: GetUsersId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '404':
          description: User not found

    put:
      summary: Update a user
      description: Update an existing user by ID
      operationId: PutUsersId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: Invalid input
        '404':
          description: User not found

    patch:
      summary: Partially update a user
      description: Partially update an existing user by ID
      operationId: PatchUsersId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: Invalid input
        '404':
          description: User not found

    delete:
      summary: Delete a user
      description: Delete a user by ID
      operationId: DeleteUsersId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          readOnly: true
          example: "1"
        name:
          type: string
          example: "Claire"

    UserCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Claire"

    UserUpdate:
      type: object
      properties:
        name:
          type: string
          example: "John Updated"