# internal/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package gen

// User defines model for User.
type User struct {
	Id   *string `json:"id,omitempty"`
	Name *string `json:"name,omitempty"`
}

// UserCreate defines model for UserCreate.
type UserCreate struct {
	Name string `json:"name"`
}

// UserUpdate defines model for UserUpdate.
type UserUpdate struct {
	Name *string `json:"name,omitempty"`
}

// PostUsersJSONRequestBody defines body for PostUsers for application/json ContentType.
type PostUsersJSONRequestBody = UserCreate

// PatchUsersIdJSONRequestBody defines body for PatchUsersId for application/json ContentType.
type PatchUsersIdJSONRequestBody = UserUpdate

// PutUsersIdJSONRequestBody defines body for PutUsersId for application/json ContentType.
type PutUsersIdJSONRequestBody = User